pyproject.toml
[project]
name = "verion-ai-grader"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "Django==5.2.6",
    "djangorestframework==3.15.2",
    "boto3==1.34.141",
    "psycopg2-binary==2.9.12",
    "dj-database-url==2.3.0",
    "gunicorn==23.0.0",
    "python-dotenv==1.0.1",
    "drf-spectacular==0.27.2",
    "drf-spectacular-sidecar==2025.5.1",
]

[project.optional-dependencies]
dev = [
    "hypothesis==6.131.15",
    "pytest==8.4.0",
    "pytest-django==4.10.0",
    "pytest-mock==3.14.0",
    "black==24.10.0",
    "isort==5.13.2",
    "flake8==7.1.1",
    "mypy==1.13.0",
    "django-stubs==5.1.1",
    "djangorestframework-stubs==3.15.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "verion_ai_grader.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "hypothesis==6.131.15",
    "pytest==8.4.0",
    "pytest-django==4.10.0",
    "pytest-mock==3.14.0",
    "black==24.10.0",
    "isort==5.13.2",
    "flake8==7.1.1",
    "mypy==1.13.0",
    "django-stubs==5.1.1",
    "djangorestframework-stubs==3.15.1",
]
